requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.21.3
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==10.0.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmudict==1.0.22
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
ctranslate2==3.24.0
cycler==0.12.1
Cython==3.0.9
decorator==5.1.1
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.110.0
faster-whisper==0.10.1
ffmpy==0.3.2
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g2p-en==2.1.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-speech==2.25.1
googleapis-common-protos==1.63.0
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.0.0
jaconv==0.3.4
jieba==0.42.1 --hash=sha256:ab43ed0da1810ff29fa5f6a513577ee4e66b256421cee16c72fb05c86667f593
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
langid==1.1.6
librosa==0.9.2
lightning==2.2.1
lightning-utilities==0.11.1
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mecab-python3==1.0.8
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.1
openai==1.14.3
optuna==3.6.0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
proces==0.1.7
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyannote.audio==3.2.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyloudnorm==0.1.1
pynndescent==0.5.11
pyopenjtalk-dict==0.3.4.dev2
pyparsing==3.1.2
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
pyworld-prebuilt==0.3.4.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
resampy==0.4.3
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.4
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
speechbrain==1.0.0
SQLAlchemy==2.0.29
srt==3.5.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.39.1
typer==0.11.0
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
uvicorn==0.29.0
vosk==0.3.44
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
yt-dlp==2024.5.27
pydub==0.25.1
moviepy==1.0.3